fix java examples for 1.11.0 release
