--- examples/java/demo-config/sdct.conf
@@ -1,18 +1,0 @@
-# absolute path to the examples/demo-project folder
-project_path=FOLDER_PATH
-
-resources_suffix=
-resources_prefix=sdct-
-# region name, example - eu-central-1
-region=REGION_NAME
-# bucket name to upload deployment artifacts, must be unique across all AWS accounts
-deploy_target_bucket=BUCKET_NAME
-# your account id
-account_id=ACCOUNT_ID
-access_role=
-
-aws_access_key_id=ACCESS_KEY_ID
-aws_secret_access_key=SECRET_ACCESS_KEY
-
-# build configuration
-build_projects_mapping=mvn:/demo-java;python:/demo-python
