--- examples/java/demo-config/sdct_aliases.conf
@@ -1,3 +1,0 @@
-region=REGION_NAME
-notification_bucket=BUCKET_NAME
-account_id=ACCOUNT_ID
