+++ examples/java/demo-java/.syndicate-config-demo-java/syndicate_aliases.yml
@@ -1,0 +1,8 @@
+# AWS account id where syndicate will deploy, example '987601234050'
+account_id: ACCOUNT_ID
+lambdas_alias_name: dev
+logs_expiration: 30
+# AWS region where syndicate will deploy, example: eu-central-1
+region: REGION_NAME
+# name of AWS S3 bucket where lambda will store notifications
+notification_bucket: NOTIFICATION_BUCKET
