+++ examples/java/demo-java/pom.xml
@@ -5,15 +5,24 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>com.aws.syndicate</groupId>
-    <artifactId>syndicate-demo</artifactId>
+    <artifactId>demo-java</artifactId>
     <version>1.0-SNAPSHOT</version>
+
+    <properties>
+        <maven-shade-plugin.version>3.2.0</maven-shade-plugin.version>
+        <syndicate.java.plugin.version>1.11.0</syndicate.java.plugin.version>
+        <maven.compiler.source>1.8</maven.compiler.source>
+        <maven.compiler.target>1.8</maven.compiler.target>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+        <src.dir>jsrc/main/java</src.dir>
+    </properties>
 
     <dependencies>
         <!--AWS dependencies-->
         <dependency>
             <groupId>com.amazonaws</groupId>
             <artifactId>aws-lambda-java-core</artifactId>
-            <version>1.1.0</version>
+            <version>1.2.3</version>
         </dependency>
         <dependency>
             <groupId>com.amazonaws</groupId>
@@ -23,35 +32,31 @@
         <dependency>
             <groupId>com.amazonaws</groupId>
             <artifactId>aws-java-sdk-s3</artifactId>
-            <version>1.11.246</version>
+            <version>1.12.659</version>
         </dependency>
         <dependency>
             <groupId>com.amazonaws</groupId>
             <artifactId>aws-java-sdk-dynamodb</artifactId>
-            <version>1.11.246</version>
+            <version>1.12.659</version>
         </dependency>
         <!--Syndicate annotations-->
         <dependency>
             <groupId>net.sf.aws-syndicate</groupId>
             <artifactId>deployment-configuration-annotations</artifactId>
-            <version>1.11.0</version>
+            <version>${syndicate.java.plugin.version}</version>
         </dependency>
     </dependencies>
 
-    <properties>
-        <maven.compiler.source>1.7</maven.compiler.source>
-        <maven.compiler.target>1.7</maven.compiler.target>
-    </properties>
-
     <build>
+        <sourceDirectory>${src.dir}</sourceDirectory>
         <plugins>
             <plugin>
                 <groupId>net.sf.aws-syndicate</groupId>
                 <artifactId>deployment-configuration-maven-plugin</artifactId>
-                <version>1.11.0</version>
+                <version>${syndicate.java.plugin.version}</version>
                 <configuration>
                     <packages>
-                        <package>com.aws.syndicate.demo</package>
+                        <package>com.demolayerurl.syndicate.demo</package>
                     </packages>
                     <fileName>${project.name}-${project.version}.jar</fileName>
                 </configuration>
@@ -69,7 +74,7 @@
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-shade-plugin</artifactId>
-                <version>3.1.0</version>
+                <version>${maven-shade-plugin.version}</version>
                 <configuration>
                     <filters>
                         <filter>
