+++ examples/java/demo-layer-url/.syndicate-config-demo-layer-url/syndicate_aliases.yml
@@ -1,0 +1,6 @@
+# AWS account id where syndicate will deploy, example '987601234050'
+account_id: ACCOUNT_ID
+lambdas_alias_name: dev
+logs_expiration: 30
+# AWS region where syndicate will deploy, example: eu-central-1
+region: REGION_NAME
