--- examples/java/demo-layer-url/jsrc/main/java/com/aws/lambda/response/ErrorMessage.java
@@ -1,8 +1,0 @@
-package com.aws.lambda.response;
-
-public class ErrorMessage {
-    private final String error;
-    public ErrorMessage(String error) {
-        this.error = error;
-    }
-}
