+++ examples/java/demo-layer-url/jsrc/main/java/com/demolayerurl/handler/HelloHandler.java
@@ -1,11 +1,27 @@
-package com.aws.lambda.handler;
+/*
+ * Copyright 2018 EPAM Systems, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.demolayerurl.handler;
 
 import com.amazonaws.services.lambda.runtime.Context;
 import com.amazonaws.services.lambda.runtime.RequestHandler;
 import com.amazonaws.services.lambda.runtime.events.APIGatewayProxyRequestEvent;
 import com.amazonaws.services.lambda.runtime.events.APIGatewayProxyResponseEvent;
-import com.aws.lambda.response.ErrorMessage;
-import com.aws.lambda.response.HelloMessage;
+import com.demolayerurl.response.ErrorMessage;
+import com.demolayerurl.response.HelloMessage;
 import com.google.gson.Gson;
 import com.syndicate.deployment.annotations.LambdaUrlConfig;
 import com.syndicate.deployment.annotations.lambda.LambdaHandler;
